final push hopefully#

+++ srcs/imports/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tamutlu <tamutlu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 13:05:34 by tamutlu           #+#    #+#             */
-/*   Updated: 2025/05/21 16:41:13 by tamutlu          ###   ########.fr       */
+/*   Updated: 2025/05/21 18:50:38 by tamutlu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,7 +66,8 @@ void				ft_putstr_fd(char *s, int fd);
 void				ft_striteri(char *s, void (*f)(unsigned int, char *));
 char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
 char				*ft_itoa(int n);
-int					ft_atoi(const char *str, int *error);
+long				ft_atoi(const char *str, int *error);
+// long				ft_atoi(const char *str);
 char				*ft_strjoin(char const *s1, char const *s2);
 void				ft_putchar(char c, int *len);
 void				ft_putstr(char *s, int *len);
